Track extension blocks in the Runtime
